DNM PoC Enable VNC console access to Ironic nodes
This is an early prototype for vnc graphical console access to Nova
instances running on Ironic baremetal nodes.

It is really hacky for now and sole aim is to test this patch
I283ddf5aec64ae4934ec3cf63ee7e357a8a899b2 with real HW.

In general though this work depends on "VNC Graphical console" spec
being approved and implemented in Ironic and its counterpart in Nova.

Change-Id: I3e98b79e01c0b3ade5622ab058101e1348716b82