Support rebuild with ephemeral encryption for qcow2
For rebuilds with a new image, the image properties are validated
against the instance's flavor to ensure there are no conflicts, else a
HTTP 400 is returned.

Rebuilds with a new image are allowed to change the instance from using
ephemeral encryption to not using ephemeral encryption via the image
properties, as long as the rebuild request is being made by the user_id
that owns the instance. This is meant to ensure encryption secrets are
always accessible to owner of the instance. Requests from different
users are rejected with HTTP 403.

Ephemeral encryption attributes in the BDMs are reset during a rebuild
from requesting ephemeral encryption to no ephemeral encryption.

Related to blueprint ephemeral-storage-encryption

Change-Id: I09113512f9e2cec442ce61ef145a8ea170cf319c
Signed-off-by: Pavlo Shchelokovskyy <shchelokovskyy@gmail.com>