Support create with ephemeral encryption for qcow2
This adds support for the qcow2 imagebackend to create encrypted
ephemeral disks for:

  * root disk created from an unencrypted source image
    (encrypted source image support is future work)
  * ephemeral disk
  * swap disk

A new unit test test_driver_capabilities_flat is also added ahead of
support for the RAW image backend. The Flat image backend was the only
backend that did not have a driver capabilities unit test.

Related to blueprint ephemeral-encryption-libvirt

Change-Id: Ie5788c791fae000573c488ea31af7a15799e1404
Signed-off-by: Pavlo Shchelokovskyy <shchelokovskyy@gmail.com>