libvirt: Add optional 'description' kwarg to create_secret()
We already have support for writing the <description> element in
LibvirtConfigSecret but we aren't exposing the ability to set a
description in the Host.create_secret() method.

This just wires things up so that a description can be provided when
creating a secret and adds the ability to parse <description> out of
secret XML as well.

The use case is to be able to set a description for libvirt secrets
that will allow us to identify whether a given secret is for ephemeral
encryption or not.

For example, if we will want to clean up unused ephemeral encryption
libvirt secrets during compute host init, we need to be sure we only
delete secrets related to ephemeral encryption and not touch Cinder
volume encryption secrets or vTPM secrets.

Related to blueprint ephemeral-encryption-libvirt

Change-Id: I99100f823b85f9b95b341a647ff4dd4927252413
Signed-off-by: Pavlo Shchelokovskyy <shchelokovskyy@gmail.com>