Add EncryptDetails object for ephemeral encryption
Whenever we create a BDM using ephemeral encryption, we want to record
what options we used when we created the disk. These options are passed
to qemu-img when creating or converting images.

We want to record this information to help support upgrade scenarios in
the future if and when QEMU encryption option defaults change or
options are added or removed. Representing the options as a versioned
object will help us keep track of changes over time and the values
could be used to support migration/conversion of images to different
cipher algorithms etc if needed.

Related to blueprint ephemeral-storage-encryption

Change-Id: I51bf4da6d669b32e8ea377f4c66099632c2a17c4
Signed-off-by: Pavlo Shchelokovskyy <shchelokovskyy@gmail.com>