Add database migration to ALTER encryption_options
The encryption_options column in the block_device_mapping table has
never been used for anything and data has never been written to it.

Now we have need to store encryption details for BDMs such as what
cipher algorithm, cipher mode, hash algorithm, etc were used to encrypt
the data. We may need these details in the future for upgrade migration
purposes.

We agreed that it would not be good to use the existing
encryption_options field on the BlockDeviceMapping object, so instead
we are adding an encryption_details field and raising an exception if
there is an attempt to set the never-to-be-used encryption_options
field.

We plan to store the encryption_details data as a serialized
EncryptDetails JSON blob which could theoretically become large in the
future, so we want to use the TEXT column type to store it.

We could either add a new encryption_details column to the
block_device_mapping table and keep the encryption_options column empty
forever or we could ALTER the column to change its name and data type
to be used for encryption_details. We are trying to do the latter to
avoid carrying an empty column which will never be used.

This adds the encryption_details field to the BlockDeviceMapping object
as an ObjectField('EncryptDetails'), raises an exception if an attempt
is made to set encryption_options, and swaps all of the code over from
encryption_options to encryption_details.

Related to blueprint ephemeral-storage-encryption

Change-Id: Iddbb31fb6c9bef1e7fba3b4dd7c6f5db2e612873
Signed-off-by: Pavlo Shchelokovskyy <shchelokovskyy@gmail.com>